backend/requirements.txt
# Flask and related packages
Flask==3.1.0
flasgger==0.9.7.1
flask_cors==5.0.0

# Tokenization and language models
accelerate==1.2.1
bitsandbytes==0.45.0
diffusers==0.32.1
langchain-core~=0.3.29
langchain-google-vertexai~=2.0.10
langchain_openai==0.2.14
Pillow>=11.1.0
tiktoken~=0.8.0
torch==2.5.1
transformers==4.47.1
openai==1.59.5
ollama==0.4.5
litellm==1.57.4

# Google Cloud Platform (GCP) related packages
google-auth~=2.37.0
google.cloud.logging~=3.11.3

# Utility libraries
python-dotenv==1.0.1
requests~=2.32.3
